# ascii heap function