--- a/pwndbg/commands/heap.py
+++ b/pwndbg/commands/heap.py
@@ -1,5 +1,6 @@
 import argparse
 import ctypes
+from string import printable
 from typing import Dict
 from typing import List
 
@@ -973,12 +974,11 @@ def vis_heap_chunks(
         )
 
 
-from string import printable
-valid_chars = list(map(ord, set(printable) - set("\t\r\n\x0c\x0b")))
+VALID_CHARS = list(map(ord, set(printable) - set("\t\r\n\x0c\x0b")))
+
 
 def bin_ascii(bs):
-    return "".join(chr(c) if c in valid_chars else "." for c in bs)
-
+    return "".join(chr(c) if c in VALID_CHARS else "." for c in bs)
 
 
 def bin_labels_mapping(collections):