benchmark before this commit
```
Fri May 12 23:14:35 2023    profile.prof

         2824849 function calls (2820117 primitive calls) in 4.677 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      2/1    0.008    0.004    4.677    4.677 {built-in method _gdb.execute}
        1    0.000    0.000    4.669    4.669 __init__.py:128(invoke)
        1    0.000    0.000    4.668    4.668 __init__.py:182(__call__)
        1    0.000    0.000    4.668    4.668 __init__.py:317(_OnlyWhenRunning)
        1    0.000    0.000    4.668    4.668 __init__.py:394(_OnlyWithResolvedHeapSyms)
        1    0.000    0.000    4.651    4.651 __init__.py:358(_try2run_heap_command)
        1    0.000    0.000    4.651    4.651 __init__.py:342(_OnlyWhenHeapIsInitialized)
        1    0.587    0.587    4.630    4.630 heap.py:825(vis_heap_chunks)
    90366    1.409    0.000    2.606    0.000 heap.py:976(bin_ascii)
    90558    0.158    0.000    1.198    0.000 {method 'join' of 'str' objects}
   813294    1.021    0.000    1.039    0.000 heap.py:980(<genexpr>)
    90368    0.158    0.000    0.598    0.000 memory.py:18(read)
    90368    0.368    0.000    0.368    0.000 {method 'read_memory' of 'gdb.Inferior' objects}
   135550    0.075    0.000    0.325    0.000 __init__.py:135(wrapper)
   135550    0.089    0.000    0.196    0.000 __init__.py:123(colorize)
    90824    0.170    0.000    0.171    0.000 {method 'format' of 'str' objects}
   135550    0.075    0.000    0.108    0.000 __init__.py:182(terminateWith)
    90366    0.057    0.000    0.098    0.000 arch.py:37(unpack)
      389    0.003    0.000    0.075    0.000 ptmalloc.py:352(__init__)
      387    0.001    0.000    0.074    0.000 ptmalloc.py:316(is_top_chunk)
      387    0.000    0.000    0.071    0.000 ptmalloc.py:309(arena)
      387    0.001    0.000    0.071    0.000 ptmalloc.py:302(heap)
    35478    0.014    0.000    0.066    0.000 __init__.py:59(yellow)
6319/6301    0.007    0.000    0.064    0.000 cache.py:120(decorator)
    33335    0.013    0.000    0.061    0.000 __init__.py:63(blue)
    32278    0.013    0.000    0.059    0.000 __init__.py:71(cyan)
    90369    0.050    0.000    0.050    0.000 {built-in method builtins.max}
        5    0.000    0.000    0.041    0.008 symbol.py:112(address)
    90366    0.041    0.000    0.041    0.000 {built-in method _struct.unpack}
        5    0.041    0.008    0.041    0.008 {built-in method _gdb.lookup_symbol}
      390    0.004    0.000    0.035    0.000 ptmalloc.py:1455(get_sbrk_heap_region)
   135556    0.033    0.000    0.033    0.000 {method 'replace' of 'str' objects}
     1162    0.003    0.000    0.032    0.000 ptmalloc.py:179(__match_renamed_field)
    17075    0.007    0.000    0.032    0.000 __init__.py:55(green)
    17384    0.007    0.000    0.031    0.000 __init__.py:67(purple)
     2324    0.005    0.000    0.029    0.000 types.py:45(has_field)
      776    0.007    0.000    0.029    0.000 ptmalloc.py:201(size)
    90756    0.028    0.000    0.028    0.000 ptmalloc.py:524(top)
        3    0.000    0.000    0.026    0.009 {built-in method builtins.print}
        8    0.000    0.000    0.026    0.003 __init__.py:52(write)
        8    0.026    0.003    0.026    0.003 {built-in method _gdb.write}
   107217    0.025    0.000    0.025    0.000 {method 'get' of 'dict' objects}
      390    0.001    0.000    0.022    0.000 ptmalloc.py:1065(get_region)
        7    0.000    0.000    0.022    0.003 typeinfo.py:92(load)
        8    0.022    0.003    0.022    0.003 {built-in method _gdb.lookup_type}
    90369    0.021    0.000    0.021    0.000 {built-in method _gdb.selected_inferior}
     2324    0.021    0.000    0.021    0.000 {method 'fields' of 'gdb.Type' objects}
        1    0.000    0.000    0.021    0.021 ptmalloc.py:1470(is_initialized)
 5070/390    0.008    0.000    0.020    0.000 copy.py:128(deepcopy)
     1166    0.006    0.000    0.018    0.000 ptmalloc.py:157(__init__)
   155895    0.018    0.000    0.018    0.000 {built-in method builtins.chr}
        2    0.000    0.000    0.017    0.009 ptmalloc.py:1314(libc_has_debug_syms)
     2359    0.006    0.000    0.016    0.000 memory.py:292(poi)
      390    0.002    0.000    0.015    0.000 copy.py:259(_reconstruct)
    90373    0.015    0.000    0.015    0.000 {method 'extend' of 'list' objects}
        1    0.000    0.000    0.015    0.015 ptmalloc.py:1089(fastbins)
      386    0.004    0.000    0.015    0.000 ptmalloc.py:211(real_size)
        2    0.000    0.000    0.014    0.007 glibc.py:177(_OnlyWhenGlibcLoaded)
      8/4    0.000    0.000    0.014    0.004 proc.py:148(wrapper)
        4    0.000    0.000    0.014    0.004 glibc.py:37(get_version)
        1    0.000    0.000    0.014    0.014 glibc.py:51(_get_version)
      387    0.001    0.000    0.014    0.000 ptmalloc.py:253(prev_inuse)
      392    0.002    0.000    0.014    0.000 ptmalloc.py:1327(main_arena)
      519    0.003    0.000    0.011    0.000 ptmalloc.py:456(__init__)
        1    0.000    0.000    0.009    0.009 ptmalloc.py:1340(thread_arena)
        1    0.001    0.001    0.009    0.009 ptmalloc.py:1118(tcachebins)
      126    0.002    0.000    0.009    0.000 ptmalloc.py:1149(bin_at)
      396    0.001    0.000    0.008    0.000 ptmalloc.py:1390(mp)
      390    0.002    0.000    0.008    0.000 copy.py:227(_deepcopy_dict)
        1    0.000    0.000    0.006    0.006 ptmalloc.py:1214(smallbins)
     2361    0.005    0.000    0.005    0.000 {method 'cast' of 'gdb.Value' objects}
        2    0.000    0.000    0.005    0.002 ptmalloc.py:600(active_heap)
        2    0.000    0.000    0.004    0.002 symbol.py:165(static_linkage_symbol_address)
        2    0.004    0.002    0.004    0.002 {built-in method _gdb.lookup_static_symbol}
      323    0.003    0.000    0.004    0.000 chain.py:29(get)
        1    0.000    0.000    0.004    0.004 ptmalloc.py:1008(multithreaded)
     2361    0.004    0.000    0.004    0.000 {method 'dereference' of 'gdb.Value' objects}
        1    0.000    0.000    0.003    0.003 ptmalloc.py:1228(largebins)
        1    0.001    0.001    0.003    0.003 heap.py:983(bin_labels_mapping)
     2324    0.002    0.000    0.003    0.000 types.py:22(get_basic_type)
      252    0.001    0.000    0.003    0.000 ptmalloc.py:1182(<lambda>)
        1    0.000    0.000    0.003    0.003 ptmalloc.py:1449(get_tcache)
        1    0.000    0.000    0.003    0.003 ptmalloc.py:1355(thread_cache)
      253    0.001    0.000    0.002    0.000 ptmalloc.py:1030(chunk_key_offset)
        1    0.000    0.000    0.002    0.002 ptmalloc.py:1425(tcache_perthread_struct)
        1    0.000    0.000    0.002    0.002 ptmalloc.py:1420(malloc_state)
        1    0.000    0.000    0.002    0.002 ptmalloc.py:1440(malloc_par)
        1    0.000    0.000    0.002    0.002 ptmalloc.py:1051(tcache_next_offset)
        1    0.000    0.000    0.002    0.002 ptmalloc.py:1430(tcache_entry)
      780    0.001    0.000    0.002    0.000 copy.py:243(_keep_alive)
      780    0.000    0.000    0.001    0.000 copy.py:264(<genexpr>)
     6710    0.001    0.000    0.001    0.000 {method 'items' of 'dict' objects}
      390    0.001    0.000    0.001    0.000 {method '__reduce_ex__' of 'object' objects}
      394    0.001    0.000    0.001    0.000 enum.py:764(__format__)
     7020    0.001    0.000    0.001    0.000 {built-in method builtins.id}
      396    0.000    0.000    0.001    0.000 memory.py:145(__contains__)
     2760    0.001    0.000    0.001    0.000 {built-in method builtins.isinstance}
        1    0.000    0.000    0.001    0.001 __init__.py:498(split_args)
      390    0.000    0.000    0.001    0.000 copyreg.py:100(__newobj__)
     2361    0.001    0.000    0.001    0.000 {method 'pointer' of 'gdb.Type' objects}
        1    0.000    0.000    0.001    0.001 argparse.py:1837(parse_args)
        1    0.000    0.000    0.001    0.001 argparse.py:1844(parse_known_args)
      776    0.001    0.000    0.001    0.000 memory.py:13(round_down)
      253    0.001    0.000    0.001    0.000 ptmalloc.py:1045(<listcomp>)
     4290    0.001    0.000    0.001    0.000 copy.py:182(_deepcopy_atomic)
        1    0.000    0.000    0.001    0.001 argparse.py:1883(_parse_known_args)
       64    0.000    0.000    0.001    0.000 ptmalloc.py:1017(_request2size)
     1175    0.000    0.000    0.000    0.000 memory.py:83(end)
        1    0.000    0.000    0.000    0.000 argparse.py:2028(consume_positionals)
        1    0.000    0.000    0.000    0.000 vmmap.py:148(find)
     2324    0.000    0.000    0.000    0.000 {method 'unqualified' of 'gdb.Type' objects}
     1748    0.000    0.000    0.000    0.000 {method 'append' of 'list' objects}      
```
after this commit:
```

Fri May 12 23:13:00 2023    xyz

         2825696 function calls (2820964 primitive calls) in 3.075 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      2/1    0.009    0.005    3.075    3.075 {built-in method _gdb.execute}
        1    0.000    0.000    3.068    3.068 __init__.py:128(invoke)
        1    0.000    0.000    3.067    3.067 __init__.py:182(__call__)
        1    0.000    0.000    3.067    3.067 __init__.py:317(_OnlyWhenRunning)
        1    0.000    0.000    3.067    3.067 __init__.py:394(_OnlyWithResolvedHeapSyms)
        1    0.000    0.000    3.049    3.049 __init__.py:358(_try2run_heap_command)
        1    0.000    0.000    3.049    3.049 __init__.py:342(_OnlyWhenHeapIsInitialized)
        1    0.503    0.503    3.025    3.025 heap.py:826(vis_heap_chunks)
    90366    0.078    0.000    1.235    0.000 heap.py:980(bin_ascii)
    90558    0.147    0.000    1.158    0.000 {method 'join' of 'str' objects}
   813294    0.989    0.000    1.010    0.000 heap.py:981(<genexpr>)
    90368    0.138    0.000    0.516    0.000 memory.py:18(read)
    90368    0.311    0.000    0.311    0.000 {method 'read_memory' of 'gdb.Inferior' objects}
   135550    0.069    0.000    0.302    0.000 __init__.py:135(wrapper)
   135550    0.086    0.000    0.185    0.000 __init__.py:123(colorize)
    90824    0.139    0.000    0.140    0.000 {method 'format' of 'str' objects}
   135550    0.070    0.000    0.099    0.000 __init__.py:182(terminateWith)
      389    0.003    0.000    0.083    0.000 ptmalloc.py:352(__init__)
    90366    0.049    0.000    0.082    0.000 arch.py:37(unpack)
      387    0.001    0.000    0.079    0.000 ptmalloc.py:316(is_top_chunk)
      387    0.000    0.000    0.075    0.000 ptmalloc.py:309(arena)
      387    0.001    0.000    0.075    0.000 ptmalloc.py:302(heap)
6319/6301    0.007    0.000    0.074    0.000 cache.py:120(decorator)
    35478    0.013    0.000    0.061    0.000 __init__.py:59(yellow)
    33335    0.012    0.000    0.057    0.000 __init__.py:63(blue)
    32278    0.012    0.000    0.056    0.000 __init__.py:71(cyan)
    90369    0.050    0.000    0.050    0.000 {built-in method builtins.max}
      390    0.004    0.000    0.042    0.000 ptmalloc.py:1455(get_sbrk_heap_region)
        5    0.001    0.000    0.042    0.008 symbol.py:112(address)
        5    0.040    0.008    0.040    0.008 {built-in method _gdb.lookup_symbol}
     1162    0.003    0.000    0.034    0.000 ptmalloc.py:179(__match_renamed_field)
    90366    0.033    0.000    0.033    0.000 {built-in method _struct.unpack}
        7    0.000    0.000    0.032    0.005 typeinfo.py:92(load)
        8    0.032    0.004    0.032    0.004 {built-in method _gdb.lookup_type}
    17075    0.006    0.000    0.031    0.000 __init__.py:55(green)
     2324    0.006    0.000    0.030    0.000 types.py:45(has_field)
   135556    0.030    0.000    0.030    0.000 {method 'replace' of 'str' objects}
      776    0.007    0.000    0.030    0.000 ptmalloc.py:201(size)
    17384    0.006    0.000    0.029    0.000 __init__.py:67(purple)
    90756    0.026    0.000    0.026    0.000 ptmalloc.py:524(top)
        3    0.000    0.000    0.025    0.008 {built-in method builtins.print}
        8    0.000    0.000    0.025    0.003 __init__.py:52(write)
        8    0.025    0.003    0.025    0.003 {built-in method _gdb.write}
        1    0.000    0.000    0.025    0.025 ptmalloc.py:1470(is_initialized)
      390    0.001    0.000    0.023    0.000 ptmalloc.py:1065(get_region)
   107217    0.023    0.000    0.023    0.000 {method 'get' of 'dict' objects}
     2324    0.022    0.000    0.022    0.000 {method 'fields' of 'gdb.Type' objects}
 5070/390    0.009    0.000    0.021    0.000 copy.py:128(deepcopy)
   156742    0.021    0.000    0.021    0.000 {built-in method builtins.chr}
     1166    0.006    0.000    0.018    0.000 ptmalloc.py:157(__init__)
        2    0.000    0.000    0.017    0.009 ptmalloc.py:1314(libc_has_debug_syms)
     2359    0.006    0.000    0.016    0.000 memory.py:292(poi)
    90369    0.016    0.000    0.016    0.000 {built-in method _gdb.selected_inferior}
      390    0.003    0.000    0.016    0.000 copy.py:259(_reconstruct)
      386    0.005    0.000    0.015    0.000 ptmalloc.py:211(real_size)
        1    0.001    0.001    0.015    0.015 ptmalloc.py:1118(tcachebins)
      387    0.001    0.000    0.015    0.000 ptmalloc.py:253(prev_inuse)
    90373    0.014    0.000    0.014    0.000 {method 'extend' of 'list' objects}
        1    0.000    0.000    0.013    0.013 ptmalloc.py:1089(fastbins)
      396    0.001    0.000    0.013    0.000 ptmalloc.py:1390(mp)
        2    0.000    0.000    0.013    0.006 glibc.py:177(_OnlyWhenGlibcLoaded)
      8/4    0.000    0.000    0.013    0.003 proc.py:148(wrapper)
        4    0.000    0.000    0.013    0.003 glibc.py:37(get_version)
      392    0.001    0.000    0.013    0.000 ptmalloc.py:1327(main_arena)
        1    0.000    0.000    0.013    0.013 glibc.py:51(_get_version)
      519    0.003    0.000    0.011    0.000 ptmalloc.py:456(__init__)
        2    0.000    0.000    0.009    0.005 ptmalloc.py:600(active_heap)
      390    0.002    0.000    0.009    0.000 copy.py:227(_deepcopy_dict)
        1    0.000    0.000    0.006    0.006 ptmalloc.py:1440(malloc_par)
      126    0.001    0.000    0.006    0.000 ptmalloc.py:1149(bin_at)
        1    0.000    0.000    0.005    0.005 ptmalloc.py:1051(tcache_next_offset)
        1    0.000    0.000    0.005    0.005 ptmalloc.py:1430(tcache_entry)
     2361    0.005    0.000    0.005    0.000 {method 'cast' of 'gdb.Value' objects}
        1    0.000    0.000    0.005    0.005 ptmalloc.py:1340(thread_arena)
        1    0.000    0.000    0.005    0.005 ptmalloc.py:1449(get_tcache)
        1    0.000    0.000    0.005    0.005 ptmalloc.py:1355(thread_cache)
      323    0.003    0.000    0.004    0.000 chain.py:29(get)
     2361    0.004    0.000    0.004    0.000 {method 'dereference' of 'gdb.Value' objects}
        1    0.000    0.000    0.004    0.004 ptmalloc.py:1425(tcache_perthread_struct)
        1    0.000    0.000    0.003    0.003 ptmalloc.py:1214(smallbins)
        1    0.000    0.000    0.003    0.003 ptmalloc.py:1228(largebins)
        2    0.000    0.000    0.003    0.002 symbol.py:165(static_linkage_symbol_address)
        2    0.003    0.002    0.003    0.002 {built-in method _gdb.lookup_static_symbol}
     2324    0.003    0.000    0.003    0.000 types.py:22(get_basic_type)
        1    0.000    0.000    0.002    0.002 ptmalloc.py:1420(malloc_state)
      252    0.001    0.000    0.002    0.000 ptmalloc.py:1182(<lambda>)
        1    0.001    0.001    0.002    0.002 heap.py:984(bin_labels_mapping)
      780    0.001    0.000    0.002    0.000 copy.py:243(_keep_alive)
        1    0.000    0.000    0.001    0.001 info.py:53(address)
      780    0.000    0.000    0.001    0.000 copy.py:264(<genexpr>)
      253    0.001    0.000    0.001    0.000 ptmalloc.py:1030(chunk_key_offset)
        1    0.000    0.000    0.001    0.001 ptmalloc.py:1008(multithreaded)
     6710    0.001    0.000    0.001    0.000 {method 'items' of 'dict' objects}
      390    0.001    0.000    0.001    0.000 {method '__reduce_ex__' of 'object' objects}
     7020    0.001    0.000    0.001    0.000 {built-in method builtins.id}
```