@disconnect3d 2023-05-12
So `vis_heap_chunks` calls the `bin_ascii` function ~90k times:
```
    90366    1.409    0.000    2.606    0.000 heap.py:976(bin_ascii)
```

And this function created the following object on each call:
```python
    valid_chars = list(map(ord, set(printable) - set("\t\r\n\x0c\x0b")))
```

This commit moved the creation of this object to a global space, so instead of creating it... 90k times we now create it just once.

This made the execution of the benchmark to take 3s instead of 4.6s, so we got it ~1.5s faster so its like ~30% faster.